chore: 쿠폰 엔티티 quantity 타입 수정
- Long => int

+++ api/src/main/java/com/coupon/api/dto/request/CreateCouponRequestDto.java
@@ -6,13 +6,13 @@ public class CreateCouponRequestDto {
     @NonNull
     private String name;
     @NonNull
-    private Long quantity;
+    private int quantity;
 
     public String getName() {
         return name;
     }
 
-    public Long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 }
